+++ R/crew_launcher.R
@@ -183,6 +183,9 @@ crew_class_launcher <- R6::R6Class(
     tls = NULL,
     #' @field processes See [crew_launcher()].
     processes = NULL,
+    #' @field profile `mirai` compute profile of the local processes,
+    #'   if applicable.
+    profile = NULL,
     #' @description Launcher constructor.
     #' @return An `R6` object with the launcher.
     #' @param name See [crew_launcher()].
@@ -270,6 +273,13 @@ crew_class_launcher <- R6::R6Class(
       }
       crew_assert(
         self$name,
+        is.character(.),
+        length(.) == 1L,
+        !anyNA(.),
+        nzchar(.)
+      )
+      crew_assert(
+        self$profile %|||% "x",
         is.character(.),
         length(.) == 1L,
         !anyNA(.),
@@ -397,6 +407,14 @@ crew_class_launcher <- R6::R6Class(
     #' @param sockets For testing purposes only.
     start = function(sockets = NULL) {
       sockets <- sockets %|||% mirai::nextget("urls", .compute = self$name)
+      if (!is.null(self$processes)) {
+        self$profile <- self$profile %|||% crew_random_name()
+        mirai::daemons(
+          n = self$processes,
+          dispatcher = FALSE,
+          .compute = self$profile
+        )
+      }
       n <- length(sockets)
       self$workers <- tibble::tibble(
         handle = replicate(n, crew_null, simplify = FALSE),
@@ -613,10 +631,55 @@ crew_class_launcher <- R6::R6Class(
       walk(x = head(x = unlaunched, n = deficit), f = self$launch)
       invisible()
     },
-    #' @description Abstract method.
-    #' @details Does not actually terminate a worker. This method is a
-    #'   placeholder, and its presence allows manual worker termination
-    #'   to be optional.
+    #' @description Run a local asynchronous task using the local
+    #'   compute profile of the launcher.
+    #' @details Used for launcher plugins with asynchronous launches and
+    #'   terminations. If `processes` is `NULL`, the task will run locally.
+    #'   Otherwise, the task will run on a local process in the local
+    #'   `mirai` compute profile.
+    #' @return If the `processes` field is `NULL`, a list with an object named
+    #'   `data` containing the result of evaluating `expr` synchronously.
+    #'   Otherwise, the task is evaluated asynchronously, and the result
+    #'   is a `mirai` task object. Either way, the `data` element
+    #'   of the return value will contain the result of the task.
+    #' @param command R code to run.
+    #' @param substitute Logical of length 1, whether to substitute `command`.
+    #'   If `FALSE`, then `command` must be an expression object
+    #'   or language object.
+    #' @param args Named list of data objects required to run `command`.
+    #' @param packages Character vector of packages to load.
+    #' @param library Character vector of library paths to load the packages
+    #'   from.
+    async = function(
+      command,
+      substitute = TRUE,
+      args,
+      packages = character(0L),
+      library = NULL
+    ) {
+      command <- if_any(substitute, substitute(command), command)
+      if_any(
+        is.null(self$processes),
+        list(
+          data = crew_async(
+            command = command,
+            args = args,
+            packages = packages,
+            library = library
+          )
+        ),
+        mirai::mirai(
+          .expr = expr_crew_async,
+          command = command,
+          args = args,
+          packages = packages,
+          library = library,
+          .compute = self$profile
+        )
+      )
+    },
+    #' @description Abstract method to terminate a single worker.
+    #' @details Launcher plugins will overwrite this method.
     #' @return `NULL` (invisibly).
     #' @param handle A handle object previously
     #'   returned by `launch_worker()` which allows the termination
@@ -647,6 +710,9 @@ crew_class_launcher <- R6::R6Class(
     #' @description Terminate the whole launcher, including all workers.
     #' @return `NULL` (invisibly).
     terminate = function() {
+      if (!is.null(self$processes)) {
+        mirai::daemons(n = 0L, .compute = self$profile)
+      }
       self$terminate_workers()
     }
   )